pyproject.toml
[build-system]
requires = [
    "flit_core >= 3.2,<4",
]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "intertitles"

[tool.flit.external-data]
directory = "data"

[tool.pyright]
typeCheckingMode = "basic"

[tool.pylint]
max-line-length = 80
disable = [
    "C0103",
    "C0114",
    "C0115",
    "C0116",
    "R0903",
    "R0913",
    "W0105",
]

[project]
name = "intertitles"
authors = [
    { name = "Jamie Hardt", email = "jamiehardt@gmail.com" },
]
readme = "README.md"
dynamic = [
    "version",
    "description",
]
requires-python = "~=3.8"
dependencies = [
    "ffmpeg-generator >= 1.0.6",
    "numpy >= 1.26.0"
]
keywords = [
    "video",
    "text",
    "ffmpeg",
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Environment :: Console",
    "Intended Audience :: Telecommunications Industry",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Graphics",
]

[project.optional-dependencies]
doc = [
    "sphinx >= 5.3.0",
    "sphinx_rtd_theme >= 1.1.1",
]

[project.urls]
Home = ""
Documentation = ""
Source = ""
Issues = ""

[project.scripts]
intertitles = "intertitles.__main__:main"
